👍Adds favorites management
