+++ iosApp/iosApp/DevFestViewModel.swift
@@ -12,12 +12,15 @@ import Combine
 import KMPNativeCoroutinesCore
 import KMPNativeCoroutinesAsync
 import NSLogger
+import SwiftUI
 
 extension Session: Identifiable { }
 
 @MainActor
 class DevFestViewModel: ObservableObject {
     let store : DevFestNantesStore
+    let defaults = UserDefaults.standard
+    var favorites: [String]
     
     @Published var venueContent: VenueContent = VenueContent(address: "5 rue de Valmy, 44000 Nantes", description: "Située en plein cœur de ville, La Cité des Congrès de Nantes propose pour le DevFest Nantes plus de 3000m² de salles de conférences, codelabs et lieu de rencontre…", latitude: 47.21308725112951, longitude: -1.542622837466317, imageUrl: "https://devfest.gdgnantes.com/static/6328df241501c6e31393e568e5c68d7e/efc43/amphi.webp", name: "Cité des Congrès de Nantes")
     @Published var agendaContent: AgendaContent = AgendaContent(sections: [])
@@ -25,8 +28,9 @@ class DevFestViewModel: ObservableObject {
     
     init() {
         self.store = DevFestNantesStoreBuilder().setUseMockServer(useMockServer: false).build()
+        self.favorites = defaults.object(forKey: "Favorites") as? [String] ?? []
     }
-
+    
     func observeSessions() async {
         do {
             let stream = asyncStream(for: store.sessionsNative)
@@ -59,7 +63,7 @@ class DevFestViewModel: ObservableObject {
     func observeVenue() async {
         Task {
             do {
-                let stream = asyncStream(for: store.getVenueNative(language: .french))
+                let stream = asyncStream(for: self.store.getVenueNative(language: .french))
                 for try await data in stream {
                     DispatchQueue.main.async {
                         self.venueContent = VenueContent(from: data)
@@ -71,65 +75,49 @@ class DevFestViewModel: ObservableObject {
                 print("Failed with error: \(error)")
             }
         }}
-
-
-        func getDate(date: String) -> Date {
-            let newFormatter = ISO8601DateFormatter()
-            return newFormatter.date(from: date) ?? Date()// replace Date String
-        }
     
-        var sectionDateFormatter: DateFormatter = {
-            let formatter = DateFormatter()
-            formatter.dateFormat = "yyyy-MM-d"
-            return formatter
-        }()
-}
-
-    struct VenueContent {
-        let address: String
-        let description: String
-        let latitude: Double
-        let longitude: Double
-        let imageUrl: String
-        let name: String
-    }
     
-    struct AgendaContent {
-        struct Session: Hashable, Identifiable {
-            let id: String
-            let abstract: String
-            let category: shared.Category?
-            let language: SessionLanguage?
-            let complexity: Complexity?
-            let openFeedbackFormId: String
-            let speakers: [Speaker]
-            let room: String
-            let date: Date
-            let startDate: String
-            let endDate: String
-            let durationAndLanguage: String
-            let title: String
-        }
-        
-        struct Section: Hashable {
-            let date: Date
-            let day: String
-            var sessions: [Session]
-        }
-        
-        var sections: [Section]
-    }
-
-private extension AgendaContent.Session {
-    
-    init(from session: Session) {
+    func getDate(date: String) -> Date {
         let newFormatter = ISO8601DateFormatter()
-        self.init(id: session.id, abstract: session.abstract, category: session.category, language: session.language, complexity: session.complexity, openFeedbackFormId: session.openFeedbackFormId, speakers: session.speakers, room: session.room?.name ?? "", date: newFormatter.date(from: session.scheduleSlot.startDate) ?? Date(), startDate: session.scheduleSlot.startDate, endDate: session.scheduleSlot.endDate,durationAndLanguage: session.getDurationAndLanguageString(), title: session.title)
+        return newFormatter.date(from: date) ?? Date()// replace Date String
+    }
+    
+    var sectionDateFormatter: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-d"
+        return formatter
+    }()
+    
+}
+
+///Favorites management
+extension DevFestViewModel {
+    
+    func toggleFavorite(ofSession session: AgendaContent.Session) {
+        if favorites.contains(session.id) {
+            self.removeSessionToFavorite(sessionId: session.id)
+        } else {
+            self.addSessionToFavorite(sessionId: session.id)
+        }
+    }
+    
+    func removeSessionToFavorite(sessionId: String) {
+        objectWillChange.send()
+        while let idx = favorites.firstIndex(of:sessionId) {
+            favorites.remove(at: idx)
+        }
+        saveFavorites()
+    }
+    
+    func addSessionToFavorite(sessionId: String) {
+        objectWillChange.send()
+        favorites.append(sessionId)
+        saveFavorites()
+    }
+    
+    func saveFavorites() {
+        defaults.set(favorites, forKey: "Favorites")
     }
 }
 
-private extension VenueContent {
-    init(from venue: Venue) {
-        self.init(address: venue.address, description: venue.description_, latitude: venue.latitude as! Double, longitude: venue.longitude as! Double, imageUrl: venue.imageUrl ?? "https://fakeface.rest/thumb/view", name: venue.name)
-    }
-}
+
